\name{frac13}
\alias{frac13}
\docType{data}
\title{
Fractional factorial
}
\description{
Fractional factorial for 13 factors
}
\usage{data("frac13")}
\format{
  A data frame with 2047 observations on the following 13 variables.
  \describe{
    \item{\code{X1}}{a numeric vector}
    \item{\code{X1.1}}{a numeric vector}
    \item{\code{X.1}}{a numeric vector}
    \item{\code{X.1.1}}{a numeric vector}
    \item{\code{X.1.2}}{a numeric vector}
    \item{\code{X.1.3}}{a numeric vector}
    \item{\code{X1.2}}{a numeric vector}
    \item{\code{X.1.4}}{a numeric vector}
    \item{\code{X.1.5}}{a numeric vector}
    \item{\code{X1.3}}{a numeric vector}
    \item{\code{X1.4}}{a numeric vector}
    \item{\code{X.1.6}}{a numeric vector}
    \item{\code{X.1.7}}{a numeric vector}
  }
}

\examples{
library(TORDs)
data(frac13)
}
\keyword{internal}
