\name{frac14}
\alias{frac14}
\docType{data}
\title{
Fractional factorial
}
\description{
Fractinal factorial of 14 factors.
}
\usage{data("frac14")}
\format{
  A data frame with 4096 observations on the following 14 variables.
  \describe{
    \item{\code{A}}{a numeric vector}
    \item{\code{B}}{a numeric vector}
    \item{\code{C}}{a numeric vector}
    \item{\code{D}}{a numeric vector}
    \item{\code{E}}{a numeric vector}
    \item{\code{F}}{a numeric vector}
    \item{\code{G}}{a numeric vector}
    \item{\code{H}}{a numeric vector}
    \item{\code{J}}{a numeric vector}
    \item{\code{K}}{a numeric vector}
    \item{\code{L}}{a numeric vector}
    \item{\code{M}}{a numeric vector}
    \item{\code{N}}{a numeric vector}
    \item{\code{O}}{a numeric vector}
  }
}

\examples{
library(TORDs)
data(frac14)

}
\keyword{internal}
